backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlmodel==0.0.14
sqlalchemy==2.0.25
pydantic-settings==2.1.0
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
psycopg2-binary==2.9.9
httpx==0.26.0
openai==1.12.0
redis==5.0.1
fakeredis==2.32.1
python-multipart==0.0.9
structlog==23.3.0
orjson==3.9.14
pymilvus==2.4.0
neo4j==5.16.0
pypdf==4.1.0
tiktoken==0.5.2
numpy==1.26.4
marshmallow==3.21.1
